feat: add extensible mempool ordering registry
